+++ src/irvine/oeis/a079/A079190.java
@@ -27,7 +27,7 @@ public class A079190 extends A001329 {
     for (final Z dd : Jaguar.factor(i).divisors()) {
       final int d = dd.intValue();
       if (((i / d) & 1) == 1) {
-        sum = sum.add(Z.valueOf(d * s[d]));
+        sum = sum.add(Z.valueOf((long) d * s[d]));
       }
     }
     return sum.pow(s[i]);
@@ -35,23 +35,23 @@ public class A079190 extends A001329 {
 
   private Z innerSum(final int[] s, final int i, final int j) {
     if (i == j) {
-      final Z sd = Integers.SINGLETON.sumdiv(i, d -> Z.valueOf(s[d] * d));
+      final Z sd = Integers.SINGLETON.sumdiv(i, d -> Z.valueOf((long) s[d] * d));
       if ((i & 1) == 1) {
-        final Z u = sd.pow(s[i] * (i * s[i] + 1) / 2);
-        final Z v = sd.subtract(1).pow(s[i] * (i * s[i] - 1) / 2);
+        final Z u = sd.pow(s[i] * ((long) i * s[i] + 1) / 2);
+        final Z v = sd.subtract(1).pow(s[i] * ((long) i * s[i] - 1) / 2);
         return u.multiply(v);
       } else {
         final Z u = even(s, i);
-        final Z v = sd.pow(i * s[i] * s[i] / 2);
-        final Z w = sd.subtract(1).pow(s[i] * (i * s[i] - 2) / 2);
+        final Z v = sd.pow((long) i * s[i] * s[i] / 2);
+        final Z w = sd.subtract(1).pow(s[i] * ((long) i * s[i] - 2) / 2);
         return u.multiply(v).multiply(w);
       }
     } else if (i < j) {
-      final Z sd = Integers.SINGLETON.sumdiv(Functions.LCM.i(i, j), d -> d < s.length ? Z.valueOf(s[d] * d) : Z.ZERO);
-      return sd.pow(Functions.GCD.i(i, j) * s[i] * s[j]);
+      final Z sd = Integers.SINGLETON.sumdiv(Functions.LCM.i(i, j), d -> d < s.length ? Z.valueOf((long) s[d] * d) : Z.ZERO);
+      return sd.pow((long) Functions.GCD.i(i, j) * s[i] * s[j]);
     } else {
-      final Z sd = Integers.SINGLETON.sumdiv(Functions.LCM.i(i, j), d -> d < s.length ? Z.valueOf(s[d] * d) : Z.ZERO);
-      return sd.subtract(1).pow(Functions.GCD.i(i, j) * s[i] * s[j]);
+      final Z sd = Integers.SINGLETON.sumdiv(Functions.LCM.i(i, j), d -> d < s.length ? Z.valueOf((long) s[d] * d) : Z.ZERO);
+      return sd.subtract(1).pow((long) Functions.GCD.i(i, j) * s[i] * s[j]);
     }
   }
