+++ src/irvine/oeis/a243/A243485.java
@@ -17,6 +17,6 @@ public class A243485 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A243485() {
-    super(1, n -> Integers.SINGLETON.sum(1, n / 2, i -> Z.valueOf(i * (n - i) * eval1(i) * eval1(n - i))));
+    super(1, n -> Integers.SINGLETON.sum(1, n / 2, i -> Z.valueOf((long) i * (n - i) * eval1(i) * eval1(n - i))));
   }
 }
