+++ src/irvine/oeis/a259/A259196.java
@@ -17,6 +17,6 @@ public class A259196 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A259196() {
-    super(12, n -> Integers.SINGLETON.sum(1, n / 6, m -> Integers.SINGLETON.sum(m, (n - m) / 5, l -> Integers.SINGLETON.sum(l, (n - l - m) / 4, k -> Integers.SINGLETON.sum(k, (n - k - l - m) / 3, j -> Integers.SINGLETON.sum(j, (n - j - k - l - m) / 2, i -> Z.valueOf(eval1(i) * eval1(j) * eval1(k) * eval1(l) * eval1(m) * eval1(n - i - j - k - l - m))))))));
+    super(12, n -> Integers.SINGLETON.sum(1, n / 6, m -> Integers.SINGLETON.sum(m, (n - m) / 5, l -> Integers.SINGLETON.sum(l, (n - l - m) / 4, k -> Integers.SINGLETON.sum(k, (n - k - l - m) / 3, j -> Integers.SINGLETON.sum(j, (n - j - k - l - m) / 2, i -> Z.valueOf((long) eval1(i) * eval1(j) * eval1(k) * eval1(l) * eval1(m) * eval1(n - i - j - k - l - m))))))));
   }
 }
