+++ src/irvine/oeis/a308/A308783.java
@@ -18,6 +18,6 @@ public class A308783 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A308783() {
-    super(0, n -> Integers.SINGLETON.sum(1, n / 4, k -> Integers.SINGLETON.sum(k, (n - k) / 3, j -> Integers.SINGLETON.sum(j, (n - j - k) / 2, i -> Z.valueOf(eval1(k) * eval1(j) * eval1(i) * eval1(n - i - j - k))))).multiply(n));
+    super(0, n -> Integers.SINGLETON.sum(1, n / 4, k -> Integers.SINGLETON.sum(k, (n - k) / 3, j -> Integers.SINGLETON.sum(j, (n - j - k) / 2, i -> Z.valueOf((long) eval1(k) * eval1(j) * eval1(i) * eval1(n - i - j - k))))).multiply(n));
   }
 }
