+++ src/irvine/oeis/a308/A308839.java
@@ -18,6 +18,6 @@ public class A308839 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A308839() {
-    super(0, n -> Integers.SINGLETON.sum(1, n / 5, l -> Integers.SINGLETON.sum(l, (n - l) / 4, k -> Integers.SINGLETON.sum(k, (n - k - l) / 3, j -> Integers.SINGLETON.sum(j, (n - j - k - l) / 2, i -> Z.valueOf(eval1(l) * eval1(k) * eval1(j) * eval1(i) * eval1(n - i - j - k - l)))))).multiply(n));
+    super(0, n -> Integers.SINGLETON.sum(1, n / 5, l -> Integers.SINGLETON.sum(l, (n - l) / 4, k -> Integers.SINGLETON.sum(k, (n - k - l) / 3, j -> Integers.SINGLETON.sum(j, (n - j - k - l) / 2, i -> Z.valueOf((long) eval1(l) * eval1(k) * eval1(j) * eval1(i) * eval1(n - i - j - k - l)))))).multiply(n));
   }
 }
