+++ src/irvine/oeis/a308/A308920.java
@@ -17,6 +17,6 @@ public class A308920 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A308920() {
-    super(0, n -> Integers.SINGLETON.sum(1, n / 6, m -> Integers.SINGLETON.sum(m, (n - m) / 5, l -> Integers.SINGLETON.sum(l, (n - l - m) / 4, k -> Integers.SINGLETON.sum(k, (n - k - l - m) / 3, j -> Integers.SINGLETON.sum(j, (n - j - k - l - m) / 2, i -> Z.valueOf(eval1(m) * eval1(l) * eval1(k) * eval1(j) * eval1(i) * eval1(n - i - k - j - l - m) * m)))))));
+    super(0, n -> Integers.SINGLETON.sum(1, n / 6, m -> Integers.SINGLETON.sum(m, (n - m) / 5, l -> Integers.SINGLETON.sum(l, (n - l - m) / 4, k -> Integers.SINGLETON.sum(k, (n - k - l - m) / 3, j -> Integers.SINGLETON.sum(j, (n - j - k - l - m) / 2, i -> Z.valueOf((long) eval1(m) * eval1(l) * eval1(k) * eval1(j) * eval1(i) * eval1(n - i - k - j - l - m) * m)))))));
   }
 }
