+++ src/irvine/oeis/a309/A309796.java
@@ -13,6 +13,6 @@ public class A309796 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A309796() {
-    super(0, n -> Integers.SINGLETON.sum(1, n / 4, k -> Integers.SINGLETON.sum(k, (n - k) / 3, j -> Integers.SINGLETON.sum(j, (n - j - k) / 2, i -> Z.valueOf(i * ((i - 1) & 1))))));
+    super(0, n -> Integers.SINGLETON.sum(1, n / 4, k -> Integers.SINGLETON.sum(k, (n - k) / 3, j -> Integers.SINGLETON.sum(j, (n - j - k) / 2, i -> Z.valueOf((long) i * ((i - 1) & 1))))));
   }
 }
