+++ src/irvine/oeis/a309/A309800.java
@@ -13,6 +13,6 @@ public class A309800 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A309800() {
-    super(0, n -> Integers.SINGLETON.sum(1, n / 4, k -> Integers.SINGLETON.sum(k, (n - k) / 3, j -> Integers.SINGLETON.sum(j, (n - j - k) / 2, i -> Z.valueOf((n - i - j - k) * ((n - i - j - k - 1) & 1))))));
+    super(0, n -> Integers.SINGLETON.sum(1, n / 4, k -> Integers.SINGLETON.sum(k, (n - k) / 3, j -> Integers.SINGLETON.sum(j, (n - j - k) / 2, i -> Z.valueOf((long) (n - i - j - k) * ((n - i - j - k - 1) & 1))))));
   }
 }
